Add Sane Defaults
By default, Ubuntu includes a defaults list for all of the applications
it expects to be installed.  Unfortunately, we aren't installing
ubuntu-desktop, and so we miss a lit of the gnome-like utilities.  As
such, a lot of the default apps are nonexistant, and so Ubuntu will
choose this first alphabetical option.  This results in text files being
opened with Abiword, and images being opened with ImageMagick Image
Viewer.

I placed this before all the apt-get installs so that, as additional
apps are installed, they may add their own mimetypes.  If we were to put
it at the very end, we would need to keep up with it after we change
every application install.

That said, if we do add programs in the future that would make good
defaults, it might not be a bad idea to edit this list.