Install Gnupg-agent for Pass integration and more
Some lounge users would like to use Pass, the standard UNIX password
manager.  Pass usage is extremely limited without a GPG-agent, including
any browser integration.  This is a complicated package that is not
easily installed in a user's home directory.

Adding this package will enable users to not only use Pass, but also
other GPG-dependant applications, including encryption in email and the
like.

I have tested it on Niblick, and confirmed that it works without any
consequences.