Fix Default Session
This commit fixes the default session on the
lightdm-webkit-greeter. Prior to these changes, if the user was not
found by the ldap-api, the user would have a blank session field, this
did not affect most users, however, it did inconvenience
administrators using the local account.  In addition, this commit
allows login to continue in the event of the utility server being
offline.  This is done by adding an onerror handler to the CORS call.